requirements.txt
python-telegram-bot>=20.0
aiohttp
pydantic
python-dotenv
loguru
aiosqlite
pydantic-settings
